generate more persist table methods
