+++ example_lite_concurrent_test.go
@@ -3,8 +3,6 @@
 
 package bart_test
 
-/*
-
 import (
 	"sync"
 	"sync/atomic"
@@ -79,4 +77,3 @@ func ExampleLite_concurrent() {
 
 	// Output:
 }
-*/
